Backend/requirements.txt
bidict==0.22.1
blinker==1.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
eventlet==0.36.1
flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pycparser==2.22
pymongo==4.5.0
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
simple-websocket==0.10.1
werkzeug==2.3.7
wsproto==1.2.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.3
